requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.10.1
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
beautifulsoup4==4.11.2
Bottleneck==1.3.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
codespell==2.2.4
colorama==0.4.6
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
dash==2.9.2
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
distlib==0.3.6
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.9.1
Flask==2.2.3
fonttools==4.39.2
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.2
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.6.3
mdurl==0.1.2
mock==5.0.1
multidict==6.0.4
mysqlclient==2.1.1
networkx==3.0
numpy==1.24.2
openai==0.27.4
opencv-python==4.7.0.72
openpyxl==3.0.10
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pandas-ta==0.3.14b0
patsy==0.5.3
Pillow==9.4.0
pipenv==2023.2.18
platformdirs==3.1.1
plotly==5.14.0
pluggy==1.0.0
pyaes==1.6.1
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.2.2
pytest-mock==3.10.0
python-dateutil==2.8.2
python-louvain==0.16
python-telegram-bot==20.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
responses==0.23.1
rfc3986==1.5.0
rich==13.3.3
rsa==4.9
scipy==1.10.1
SecretStorage==3.3.3
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.4
statsmodels==0.13.5
Telethon==1.26.1
tenacity==8.2.2
tqdm==4.64.1
tradingview-ta==3.3.0
trio==0.22.0
trio-websocket==0.10.0
types-PyYAML==6.0.12.9
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
virtualenv==20.21.0
virtualenv-clone==0.5.7
websocket-client==1.5.1
websockets==11.0
Werkzeug==2.2.3
wsproto==1.2.0
yarl==1.8.2
